requirements.txt
apiview==1.3.25
appdirs==1.4.3
asgiref==3.3.1
beautifulsoup4==4.9.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-cors-headers==3.7.0
django-import-export==2.5.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
lxml==4.6.2
MarkupPy==1.14
msgpack==0.6.2
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.3
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
soupsieve==2.2
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.25.8
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
